pyproject.toml
[project]
name = "momorice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord>=2.3.2",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
]
